AI/chatbot/requirements.txt
langchain_community
langchain
langchain_core
langchain_openai
pypdf
langchain-milvus
pymilvus[milvus_lite]
langchain-ollama
gradio
python-dotenv
autogen-agentchat
autogen-core
autogen-ext